+++ src/Pim/Bundle/CatalogBundle/Manager/ProductManager.php
@@ -59,9 +59,6 @@ class ProductManager implements SaverInterface, BulkSaverInterface, RemoverInter
     /** @var AttributeRepository */
     protected $attributeRepository;
 
-    /** @var AttributeOptionRepository */
-    protected $attOptionRepository;
-
     /**
      * Constructor
      *
@@ -74,7 +71,6 @@ class ProductManager implements SaverInterface, BulkSaverInterface, RemoverInter
      * @param ProductRepositoryInterface $productRepository
      * @param AssociationTypeRepository  $assocTypeRepository
      * @param AttributeRepository        $attributeRepository
-     * @param AttributeOptionRepository  $attOptionRepository
      */
     public function __construct(
         $configuration,
@@ -85,8 +81,7 @@ class ProductManager implements SaverInterface, BulkSaverInterface, RemoverInter
         ProductBuilder $builder,
         ProductRepositoryInterface $productRepository,
         AssociationTypeRepository $assocTypeRepository,
-        AttributeRepository $attributeRepository,
-        AttributeOptionRepository $attOptionRepository
+        AttributeRepository $attributeRepository
     ) {
         $this->configuration = $configuration;
         $this->productSaver = $productSaver;
@@ -97,7 +92,6 @@ class ProductManager implements SaverInterface, BulkSaverInterface, RemoverInter
         $this->productRepository = $productRepository;
         $this->assocTypeRepository = $assocTypeRepository;
         $this->attributeRepository = $attributeRepository;
-        $this->attOptionRepository = $attOptionRepository;
     }
 
     /**
@@ -415,16 +409,6 @@ class ProductManager implements SaverInterface, BulkSaverInterface, RemoverInter
     public function getAttributeRepository()
     {
         return $this->attributeRepository;
-    }
-
-    /**
-     * Return related repository
-     *
-     * @return ObjectRepository
-     */
-    public function getAttributeOptionRepository()
-    {
-        return $this->attOptionRepository;
     }
 
     /**